+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.JYPAndroid">
         <activity
-            android:name="com.jyp.jypandroid.MainActivity"
+            android:name="com.jyp.jypandroid.SplashActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.JYPAndroid">
