+++ app/src/main/java/com/jyp/jypandroid/SplashScreen.kt
@@ -1,0 +1,27 @@
+package com.jyp.jypandroid
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun SplashScreen() {
+    Box(
+            modifier = Modifier.fillMaxSize(),
+    ) {
+        Text(
+                modifier = Modifier.align(Alignment.Center),
+                text = "splash",
+        )
+    }
+}
+
+@Composable
+@Preview(showBackground = true)
+fun SplashScreenPreview() {
+    SplashScreen()
+}
